Univariatize even more.
